// src/open-api/model/groupRole.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type GroupRole = 'OWNER' | 'VIEWER' | 'EDITOR';

export const GroupRole = {
    Owner: 'OWNER' as GroupRole,
    Viewer: 'VIEWER' as GroupRole,
    Editor: 'EDITOR' as GroupRole
};