// src/open-api/model/itemStatus.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ItemStatus = 'OPEN' | 'RESOLVED' | 'DRAFT';

export const ItemStatus = {
    Open: 'OPEN' as ItemStatus,
    Resolved: 'RESOLVED' as ItemStatus,
    Draft: 'DRAFT' as ItemStatus
};